Social Managing App.
Is a web application for managing users and group of users. Project was built under Spring Model View Controller principles and Maven dependencies. For backend processes is responsible Java Object Orietned Programming Language under Spring framework. User interface was built with help of HTML, CSS, JQuery and Bootstrap library. With help of Thymeleaf data is extracted and shown on a website. MySQL database was build under Spring Data, Hibernate technologies. On the top of it REST api was created to let to talk external enviroments with the database of the app  The main requirements of the app are following points:

Users:
1. Display list of users
2. Add user
3. Edit user
4. Delete user
5. Assign user to appropriate group
6. Remove user from appropriate group

Groups:
1. Display list of groups
2. Add group
3. Edit group
4. Delete group
5. Add user to group
6. Remove user from group